+++ Contents/mods/Evolving Traits World - Trait Sandbox/42.19/media/lua/client/ETWTS_RegisterTraitOptions.lua
@@ -69,6 +69,7 @@ local traits = {
 	{ id = "FurnitureAssembler", pointValue = -4 },
 	{ id = "GunEnthusiast", pointValue = -6 },
 	{ id = "GymRat", pointValue = -8 },
+	{ id = "CouchPotato", pointValue = 8 },
 	{ id = "Hoarder", pointValue = -4 },
 	{ id = "HomeCook", pointValue = -2 },
 	{ id = "Homichlophobia", pointValue = 1, toggle = "FogTraitsEnabled" },